refactor: Componentizando os elementos
